+++ pygpseq/anim/analysis.py
@@ -72,6 +72,7 @@ class Analyzer(iot.IOinterface):
     nbins = 200
     do_clear_Z_borders = False
     do_fill_holes = True
+    center_as_percentile = False
     rescale_deconvolved = False
     correctCA = False
     normalize_distance = True