+++ pygpseq/anim/nucleus.py
@@ -354,16 +354,22 @@ class Nucleus(iot.IOinterface):
 			sig = sig[mid, :, :]
 
 		# Perform distance transform
-		if 3 == mask.shape:
+		if 3 == len(mask.shape):
 			zero_slice = np.zeros((1, mask.shape[1], mask.shape[2]))
 			laminD = distance_transform_edt(
-				np.stack(zero_slice, mask, zero_slice),
+				np.concatenate([zero_slice, mask, zero_slice]),
 				aspect[3-len(mask.shape):])[1:-1, :, :]
 		else:
 			laminD = distance_transform_edt(mask,
 				aspect[3-len(mask.shape):])
-		centrD = distance_transform_edt(laminD != laminD.max(),
-			aspect[3-len(mask.shape):])
+
+		if kwargs['center_as_percentile']:
+			centrD = distance_transform_edt(
+				laminD < np.percentile(laminD, 99.),
+				aspect[3-len(mask.shape):])
+		else:
+			centrD = distance_transform_edt(laminD != laminD.max(),
+				aspect[3-len(mask.shape):])
 
 		# Export single-nucleus images in debugging mode
 		if debugging: